assets/scripts/requirements.txt
discord.py==1.7.3
python-dotenv==0.19.2
vgamepad==0.0.5
XInput==0.1.3